Refactor #349 - move compose-ext to ui-compose
