+++ core/component/tag/src/main/kotlin/com/nlab/reminder/core/component/tag/edit/ui/compose/TagEditStateHandler.kt
@@ -44,7 +44,6 @@ fun TagEditStateHandler(
         is TagEditState.Intro -> {
             TagEditIntroDialog(
                 tagName = state.tag.name,
-                usageCount = state.usageCount,
                 onDismissRequest = onCompleted,
                 onRenameRequestClicked = onRenameRequestClicked,
                 onDeleteRequestClicked = onDeleteRequestClicked
