+++ core/ui-compose/src/main/AndroidManifest.xml
