+++ core/ui-compose/src/main/kotlin/com/nlab/reminder/core/ui/compose/ColorPressButton.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.androidx.compose.ui
+package com.nlab.reminder.core.ui.compose
 
 import android.content.res.Configuration.UI_MODE_NIGHT_NO
 import androidx.compose.foundation.clickable
