+++ core/ui-compose/src/main/kotlin/com/nlab/reminder/core/ui/compose/TextFieldValues.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.androidx.compose.ui
+package com.nlab.reminder.core.ui.compose
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
